refactor: static 파일 읽기 기능을 위한 구조 분리 - Controller, AbstractController 분리
