+++ src/main/java/webserver/controller/Controller.java
@@ -5,6 +5,4 @@ import webserver.http.response.HttpResponse;
 
 public interface Controller {
 	HttpResponse process(HttpRequest httpRequest);
-	HttpResponse doGet(HttpRequest httpRequest);
-	HttpResponse doPost(HttpRequest httpRequest);
 }
