+++ src/main/java/webserver/controller/IndexController.java
@@ -1,30 +1,15 @@
 package webserver.controller;
 
-import webserver.http.HttpMethod;
 import webserver.http.request.HttpRequest;
 import webserver.http.response.HttpResponse;
-import webserver.http.response.HttpStatus;
-import webserver.http.response.Resource;
 
-public class IndexController implements Controller {
-	@Override
-	public HttpResponse process(HttpRequest httpRequest) {
-		if(HttpMethod.isGet(httpRequest.getMethod())) {
-			return doGet(httpRequest);
-		}
-		return doPost(httpRequest);
-	}
+public class IndexController extends AbstractController {
+
+	private static final String RESOURCE_PATH = "./templates";
 
 	@Override
-	public HttpResponse doGet(HttpRequest httpRequest) {
-		Resource resource = Resource.of("./templates" + httpRequest.getPath());
-
-		return new HttpResponse.Builder()
-				.statusLine(httpRequest.getProtocol(), HttpStatus.OK)
-				.contentType(resource.getContentType())
-				.contentLength(resource.getContentLength())
-				.responseBody(resource)
-				.build();
+	protected String getResourcePath() {
+		return RESOURCE_PATH;
 	}
 
 	@Override
