+++ src/main/java/webserver/controller/StaticController.java
@@ -1,0 +1,11 @@
+package webserver.controller;
+
+import webserver.http.request.HttpRequest;
+import webserver.http.response.HttpResponse;
+
+public class StaticController extends AbstractController {
+	@Override
+	HttpResponse doPost(HttpRequest httpRequest) {
+		return null;
+	}
+}
